Move repofiles from modules to services
