+++ integrations/repofiles_delete_test.go
@@ -10,9 +10,9 @@ import (
 
 	"code.gitea.io/gitea/models"
 	"code.gitea.io/gitea/models/unittest"
-	"code.gitea.io/gitea/modules/repofiles"
 	api "code.gitea.io/gitea/modules/structs"
 	"code.gitea.io/gitea/modules/test"
+	"code.gitea.io/gitea/services/repository/repofiles"
 
 	"github.com/stretchr/testify/assert"
 )
