+++ routers/api/v1/repo/file.go
@@ -15,11 +15,11 @@ import (
 	"code.gitea.io/gitea/models/unit"
 	"code.gitea.io/gitea/modules/context"
 	"code.gitea.io/gitea/modules/git"
-	"code.gitea.io/gitea/modules/repofiles"
 	api "code.gitea.io/gitea/modules/structs"
 	"code.gitea.io/gitea/modules/web"
 	"code.gitea.io/gitea/routers/common"
 	"code.gitea.io/gitea/routers/web/repo"
+	"code.gitea.io/gitea/services/repository/repofiles"
 )
 
 // GetRawFile get a file by path on a repository
