+++ routers/api/v1/repo/tree.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 
 	"code.gitea.io/gitea/modules/context"
-	"code.gitea.io/gitea/modules/repofiles"
+	"code.gitea.io/gitea/services/repository/repofiles"
 )
 
 // GetTree get the tree of a repository.
