+++ services/cron/tasks_basic.go
@@ -10,11 +10,11 @@ import (
 
 	"code.gitea.io/gitea/models"
 	"code.gitea.io/gitea/models/webhook"
-	repository_service "code.gitea.io/gitea/modules/repository"
 	"code.gitea.io/gitea/modules/setting"
 	"code.gitea.io/gitea/services/auth"
 	"code.gitea.io/gitea/services/migrations"
 	mirror_service "code.gitea.io/gitea/services/mirror"
+	repository_service "code.gitea.io/gitea/services/repository"
 )
 
 func registerUpdateMirrorTask() {
