+++ services/repository/repofiles/file.go
@@ -7,6 +7,7 @@ package repofiles
 import (
 	"fmt"
 	"net/url"
+	"path"
 	"strings"
 	"time"
 
@@ -122,4 +123,17 @@ func GetAuthorAndCommitterUsers(author, committer *IdentityOptions, doer *models
 		committerUser = authorUser // No valid committer so use the author as the committer (was set to a valid user above)
 	}
 	return authorUser, committerUser
+}
+
+// CleanUploadFileName Trims a filename and returns empty string if it is a .git directory
+func CleanUploadFileName(name string) string {
+	// Rebase the filename
+	name = strings.Trim(path.Clean("/"+name), " /")
+	// Git disallows any filenames to have a .git directory in them.
+	for _, part := range strings.Split(name, "/") {
+		if strings.ToLower(part) == ".git" {
+			return ""
+		}
+	}
+	return name
 }
